requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
pytest
aiosqlite
